1.10-完成加速度输出格式转换，修改光流imu融合代码，未测试

+++ src/main/telemetry/mavlink.h
@@ -20,7 +20,7 @@
 
 #pragma once
 
-extern uint16_t scale1;
+extern float scale1;
 
 void initMAVLinkTelemetry(void);
 void handleMAVLinkTelemetry(void);
